First step in replacing easysnmp as a dependency.
The goal is to have all SNMP queries done by calling the net-snmp
command-line binaries as subprocesses. It's drastically slower than
easysnmp, but infinitely more portable. To get around the slowness,
I aim to replace all the individual get requests for interface
properties with a single snmptable request for the entire ifTable
and ifXTable, and replace the individual get requests for system
properties with an snmpgetbulk request